fix: fix render added new note in end of list
