+++ src/modules/helpers/checkInputFields.js
@@ -1,0 +1,13 @@
+function checkInputFields() {
+  let msgError = '';
+  const form = document.forms.note;
+  if (!form.name.value.trim()) {
+    msgError += 'Name is empty\n';
+  }
+
+  if (!form.category.selectedIndex) {
+    msgError += 'Not selected category';
+  }
+
+  return msgError;
+}
